Change Origin header on proxied WebSocket requests
Similar to CORS controls, WebSocket servers are often (see below)
configured to check the Origin header of incoming requests. The config
for the HTTP proxy overwrites requests' Origin header to prevent issues
with CORS, but didn't previously do the same thing for WebSockets, as
they are controlled by a different config key.

Fixes #10878